Create api package and api.model package
Move .manager package to api;
Move .dto and .entity packages to api.model;
Move .repository package to api;
Move .rest package to api
