+++ src/main/java/com/CezaryZal/api/model/dto/DimensionDTO.java
@@ -1,4 +1,4 @@
-package com.CezaryZal.dto;
+package com.CezaryZal.api.model.dto;
 
 public interface DimensionDTO {
