+++ src/main/java/com/CezaryZal/api/repository/BasicDeviationsRepoDbImpl.java
@@ -1,33 +1,37 @@
-package com.CezaryZal.repository;
+package com.CezaryZal.api.repository;
 
-import com.CezaryZal.entity.NominalTolerance;
+import com.CezaryZal.api.model.entity.BasicDeviations;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
 
-public class NominalToleranceRepoDbImpl implements RepositoryDb{
+public class BasicDeviationsRepoDbImpl implements RepositoryDb {
 
     private final SessionFactory sessionFactory;
 
     @Autowired
-    public NominalToleranceRepoDbImpl(SessionFactory sessionFactory) {
+    public BasicDeviationsRepoDbImpl(SessionFactory sessionFactory) {
         this.sessionFactory = sessionFactory;
     }
 
     @Override
-    public NominalTolerance getRecord(String inputSign, int inputValue) {
+    public BasicDeviations getRecord(String inputSign, int inputValue){
 
         Session currentSession = sessionFactory.getCurrentSession();
 
-        Query<NominalTolerance> query = currentSession.createQuery("FROM NominalTolerance WHERE sign=:name " +
+        Query<BasicDeviations> query = currentSession.createQuery("FROM BasicDeviations WHERE sign=:name " +
                 "AND :value BETWEEN nominal_dimension_min and nominal_dimension_max");
         query.setParameter("name", inputSign);
         query.setParameter("value", inputValue);
 
         return query.getSingleResult();
     }
+
 }
+
+
