+++ src/main/java/com/CezaryZal/api/repository/NominalToleranceRepoDbImpl.java
@@ -1,37 +1,33 @@
-package com.CezaryZal.repository;
+package com.CezaryZal.api.repository;
 
-import com.CezaryZal.entity.BasicDeviations;
+import com.CezaryZal.api.model.entity.NominalTolerance;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
 
-public class BasicDeviationsRepoDbImpl implements RepositoryDb {
+public class NominalToleranceRepoDbImpl implements RepositoryDb{
 
     private final SessionFactory sessionFactory;
 
     @Autowired
-    public BasicDeviationsRepoDbImpl(SessionFactory sessionFactory) {
+    public NominalToleranceRepoDbImpl(SessionFactory sessionFactory) {
         this.sessionFactory = sessionFactory;
     }
 
     @Override
-    public BasicDeviations getRecord(String inputSign, int inputValue){
+    public NominalTolerance getRecord(String inputSign, int inputValue) {
 
         Session currentSession = sessionFactory.getCurrentSession();
 
-        Query<BasicDeviations> query = currentSession.createQuery("FROM BasicDeviations WHERE sign=:name " +
+        Query<NominalTolerance> query = currentSession.createQuery("FROM NominalTolerance WHERE sign=:name " +
                 "AND :value BETWEEN nominal_dimension_min and nominal_dimension_max");
         query.setParameter("name", inputSign);
         query.setParameter("value", inputValue);
 
         return query.getSingleResult();
     }
-
 }
-
-
