+++ src/main/java/com/CezaryZal/api/rest/HomeController.java
@@ -1,4 +1,4 @@
-package com.CezaryZal.rest;
+package com.CezaryZal.api.rest;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
