+++ src/main/java/com/CezaryZal/api/rest/ScoreRestController.java
@@ -1,7 +1,7 @@
-package com.CezaryZal.rest;
+package com.CezaryZal.api.rest;
 
-import com.CezaryZal.dto.DimensionDTOImpl;
-import com.CezaryZal.manager.DimensionService;
+import com.CezaryZal.api.model.dto.DimensionDTOImpl;
+import com.CezaryZal.api.manager.DimensionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
